Refactor: Introduce common superclass of Consecutive and SeparatedStorageCanvas
